requirements.txt
pycryptodomex
pyperclip
python-tk
